status: Add IgnoreWarning flag
This flag introduces a new default behavior in the status package which
will treat warnings as a non-ready state when `Wait=true`. By default
`IgnoreWarnings` is `false`, meaning that if `Wait=true`, then we will
not indicate readiness until all warnings are resolved (or the wait
timeout is hit).

If users need more fine-grained control over which pods must be warning
free and which not, they can use the pre-existing `WarnFreePods`, which
will takes precendence over the new flag.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>